change 1. memory_detection to use osal  functions 2. event to use osal  functions 3. editing work_queue to use osal functions but not sucessfully 4. add SEGGER_RTT files for debug output

+++ Usr/port/osal/port/freertos/osal_memory_port.c
@@ -48,8 +48,8 @@ void* osal_port_freertos_memory_calloc(uint32_t num, uint32_t size) {
     return ptr;
 }
 
-uint32_t osal_port_freertos_memory_get_free_size(void) { return xPortGetFreeHeapSize(); }
+size_t osal_port_freertos_memory_get_free_size(void) { return xPortGetFreeHeapSize(); }
 
-uint32_t osal_port_freertos_memory_get_minimum_free_size(void) {
+size_t osal_port_freertos_memory_get_minimum_free_size(void) {
     return xPortGetMinimumEverFreeHeapSize();
 }
